% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATmv.R
\name{MATmv}
\alias{MATmv}
\title{MATmv: Selección y análisis de datos faltantes}
\usage{
MATmv(dataframe, columnas)
}
\arguments{
\item{dataframe}{Un dataframe.}

\item{columnas}{Un vector de nombres de columnas a seleccionar.}
}
\value{
El nombre del nuevo dataframe con casos completos.
}
\description{
Esta función selecciona columnas específicas de un dataframe, crea un nuevo dataframe con casos completos y genera información sobre los datos faltantes.
}
